The gallery has been moved to a separate file, the config loader has been moved out of core (since it is only for drawing logic), misleading variables have been renamed
