+++ Brushshe/assets/palettes/seven_palette.hex
